Remove Wala and Base analyzer. Make opal independent from base analyzer
