+++ analyzer/javacg-opal/pom.xml
@@ -22,12 +22,6 @@
             <!--<scope>provided</scope>-->
         </dependency>
         <dependency>
-            <groupId>eu.fasten.analyzer</groupId>
-            <artifactId>base-analyzer</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-<!--            <scope>provided</scope>-->
-        </dependency>
-        <dependency>
             <groupId>de.opal-project</groupId>
             <artifactId>bytecode-representation_2.12</artifactId>
             <version>1.0.0</version>
@@ -57,6 +51,16 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <version>1.7.30</version>
+        </dependency>
+        <dependency>
+            <groupId>org.dom4j</groupId>
+            <artifactId>dom4j</artifactId>
+            <version>2.1.3</version>
+        </dependency>
+        <dependency>
+            <groupId>jaxen</groupId>
+            <artifactId>jaxen</artifactId>
+            <version>1.2.0</version>
         </dependency>
     </dependencies>
