+++ analyzer/javacg-opal/src/main/java/eu/fasten/analyzer/javacgopal/data/MavenCoordinate.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package eu.fasten.analyzer.baseanalyzer;
+package eu.fasten.analyzer.javacgopal.data;
 
 import eu.fasten.core.data.RevisionCallGraph;
 import java.io.BufferedReader;
@@ -30,7 +30,6 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
