+++ analyzer/javacg-opal/src/test/java/eu/fasten/analyzer/javacgopal/data/callgraph/PartialCallGraphTest.java
@@ -21,7 +21,7 @@ package eu.fasten.analyzer.javacgopal.data.callgraph;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import eu.fasten.analyzer.baseanalyzer.MavenCoordinate;
+import eu.fasten.analyzer.javacgopal.data.MavenCoordinate;
 import eu.fasten.analyzer.javacgopal.data.OPALType;
 import eu.fasten.analyzer.javacgopal.data.PartialCallGraph;
 import eu.fasten.core.data.RevisionCallGraph;
