+++ analyzer/javacg-opal/src/test/java/eu/fasten/analyzer/javacgopal/merge/CallGraphUtilsTest.java
@@ -18,7 +18,7 @@
 
 package eu.fasten.analyzer.javacgopal.merge;
 
-import eu.fasten.analyzer.baseanalyzer.MavenCoordinate;
+import eu.fasten.analyzer.javacgopal.data.MavenCoordinate;
 import eu.fasten.core.data.RevisionCallGraph;
 import eu.fasten.analyzer.javacgopal.data.PartialCallGraph;
 import java.io.File;
