--- analyzer/javacg-wala/src/main/resources/Java60RegressionExclusions.txt
@@ -1,19 +1,0 @@
-java\/awt\/.*
-javax\/swing\/.*
-sun\/awt\/.*
-sun\/swing\/.*
-com\/sun\/.*
-sun\/.*
-org\/netbeans\/.*
-org\/openide\/.*
-com\/ibm\/crypto\/.*
-com\/ibm\/security\/.*
-org\/apache\/xerces\/.*
-dalvik\/.*
-java\/io\/ObjectStreamClass*
-apple\/.*
-com\/apple\/.*
-jdk\/.*
-org\/omg\/.*
-org\/w3c\/.*
-javax\/.*
