--- analyzer/javacg-wala/src/test/resources/Java60RegressionExclusions.txt
@@ -1,22 +1,0 @@
-java\/awt\/.*
-javax\/swing\/.*
-sun\/awt\/.*
-sun\/swing\/.*
-com\/sun\/.*
-sun\/.*
-org\/netbeans\/.*
-org\/openide\/.*
-com\/ibm\/crypto\/.*
-com\/ibm\/security\/.*
-org\/apache\/xerces\/.*
-dalvik\/.*
-java\/io\/ObjectStreamClass*
-apple\/.*
-com\/apple\/.*
-jdk\/.*
-org\/omg\/.*
-org\/w3c\/.*
-javax\/.*
-java.net\/.*
-java.io\/.*
-java.nio\/.*
