+++ analyzer/pom.xml
@@ -24,12 +24,10 @@
         </dependency>
     </dependencies>
     <modules>
-        <module>javacg-wala</module>
         <module>javacg-opal</module>
         <module>dummy-analyzer</module>
         <module>metadata-plugin</module>
         <module>graph-plugin</module>
-        <module>base-analyzer</module>
         <module>repo-cloner-plugin</module>
     </modules>
